+++ CHANGES.rst
@@ -11,6 +11,10 @@ Version 3.3
 
 Unreleased
 
+* :mod:`~reader._plugins.twitter` plugin:
+  don't fail when deserializing tweets with missing ``edit_history_tweet_ids``
+  (fails in tweepy 4.11, warns in tweepy >4.12).
+
 
 Version 3.2
 -----------